pyproject.toml
[tool.poetry]
name = "venn"
version = "0.1.0"
description = ""
authors = ["Yoni Jacobi <yoni-j@drorisrael.org.il>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
sodapy = "^2.1.0"
Flask-SQLAlchemy = "^2.5.1"
Flask-RESTful = "^0.3.9"
Flask-Cors = "^3.0.10"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.26.1"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
